Rename new methods, use ffi package name.
