+++ cdm-test/src/test/java/ucar/nc2/ft/point/writer2/TestCFPointWriter.java
@@ -27,7 +27,7 @@ import ucar.nc2.ft.FeatureDataset;
 import ucar.nc2.ft.FeatureDatasetFactoryManager;
 import ucar.nc2.ft.FeatureDatasetPoint;
 import ucar.nc2.ft.point.TestCFPointDatasets;
-import ucar.nc2.jni.NetcdfClibrary;
+import ucar.nc2.ffi.netcdf.NetcdfClibrary;
 import ucar.nc2.util.CompareNetcdf2;
 import ucar.nc2.write.NetcdfFileFormat;
 import ucar.unidata.util.test.CheckPointFeatureDataset;
@@ -89,7 +89,7 @@ public class TestCFPointWriter {
   @Test
   public void testWrite4classic() throws IOException {
     // Ignore this test if NetCDF-4 isn't present.
-    if (!NetcdfClibrary.isClibraryPresent()) {
+    if (!NetcdfClibrary.isLibraryPresent()) {
       return;
     }
 
@@ -103,7 +103,7 @@ public class TestCFPointWriter {
   @Ignore("doesnt work: coordinate axes that are members of nc4 structures")
   public void testWrite4() throws IOException {
     // Ignore this test if NetCDF-4 isn't present.
-    if (!NetcdfClibrary.isClibraryPresent()) {
+    if (!NetcdfClibrary.isLibraryPresent()) {
       return;
     }
