+++ netcdf4/src/main/java/ucar/nc2/ffi/netcdf/package.html
@@ -4,11 +4,11 @@
   -->
 <HTML>
 <HEAD>
-  <TITLE>package ucar.nc2.jni</TITLE>
+  <TITLE>package ucar.nc2.ffi.netcdf</TITLE>
 </HEAD>
 <BODY>
 
-Public interface to the NetCDF C library.
+Foreign function interface to the NetCDF C library.
 
 </BODY>
 </HTML>
