+++ netcdf4/src/main/java/ucar/nc2/jni/netcdf/Nc4wrapper.java
@@ -22,7 +22,7 @@ import com.sun.jna.ptr.IntByReference;
  */
 @Deprecated
 public class Nc4wrapper implements Nc4prototypes {
-  private static boolean TRACE;
+  public static boolean TRACE;
 
   private static org.slf4j.Logger log = org.slf4j.LoggerFactory.getLogger(Nc4wrapper.class);
